+++ src/language/WhileLine.kt
@@ -1,0 +1,6 @@
+package language
+
+class WhileLine(lineNo: Int, val condition: String): Line(lineNo) {
+
+    override fun pseudocode(): String = "while $condition"
+}