+++ Changelog.md
@@ -7,6 +7,7 @@ In development
 * Objects now contain fields which can be referenced and used (but not reassigned) within your code using the syntax `object's field`, for example: `player's position` is a vector equal to the player's current position in the world
 * Added several new data types: boolean, integer, decimal, string, vector and world (note that most of these do not yet have all of the desired features, such as an easy way to perform arithmetic)
 * Added boolean expressions, which are written with very English-y syntax, e.g. `a is equal to b` or `a is less than b and a is greater than c`
+* Added if statements, which use the syntax `if condition then ... end if`
 * All block behaviours now have access to the world
 * Added an attribute class to store the value and data type of the different attributes for items, blocks, etc
 * Lines starting with invalid words will now throw an error